+++ test/mock/normal/src/http/get-get_p_c-000param-catchall/index.js
@@ -1,6 +1,6 @@
 exports.handler = async (event) => {
   const body = event
-  body.message = 'Hello from get /path/* running the default runtime'
+  body.message = 'Hello from any /get-p-c/:param/*'
   return {
     statusCode: 200,
     headers: { 'content-type': 'application/json' },